refactor: move cloud related files to cloud_storage directory
