+++ app/cloud_storage/cloud_storage_interface.py
